Done for the night - started moving spire.doc to it's own module and writing tests for it. Did some other house keeping stuff
