+++ scratches/scratch_5.py
@@ -10,6 +10,19 @@ Assumptions:
     - create embeddings per each permutation of above
 """
 
+import os, sys, tempfile, pathlib
+
+project_root = os.path.realpath(os.path.join(os.path.dirname(__file__), '../../..'))
+sys.path.insert(0, os.path.abspath(project_root))
+
+
+list_of_sentence = ["Sample sentence 1 plus some padding",
+                    "Sample sentence 2 less padding",
+                    "Sample sentence 3 all the padding you want",
+                    "Sample sentence 3 nothing"]
+
+
+
 
 
 def main():
